Deselect blocks on click outside
