+++ editor/modes/visual-editor/index.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { connect } from 'react-redux';
-
-/**
  * WordPress dependencies
  */
 import { __ } from 'i18n';
@@ -13,114 +8,19 @@ import { __ } from 'i18n';
  */
 import './style.scss';
 import Inserter from '../../inserter';
-import VisualEditorBlock from './block';
+import VisualEditorBlockList from './block-list';
 import PostTitle from '../../post-title';
-import {
-	getBlockUids,
-	getBlockInsertionPoint,
-	getBlockSelectionStart,
-	getBlockSelectionEnd,
-} from '../../selectors';
 
-const INSERTION_POINT_PLACEHOLDER = '[[insertion-point]]';
-
-class VisualEditor extends wp.element.Component {
-	constructor( props ) {
-		super( props );
-
-		this.onSelectionStart = this.onSelectionStart.bind( this );
-		this.onSelectionChange = this.onSelectionChange.bind( this );
-		this.onSelectionEnd = this.onSelectionEnd.bind( this );
-
-		this.state = {
-			selectionAtStart: null,
-		};
-	}
-
-	onSelectionStart( uid ) {
-		this.setState( { selectionAtStart: uid } );
-	}
-
-	onSelectionChange( uid ) {
-		const { onMultiSelect, selectionStart, selectionEnd } = this.props;
-		const { selectionAtStart } = this.state;
-		const isAtStart = selectionAtStart === uid;
-
-		if ( ! selectionAtStart ) {
-			return;
-		}
-
-		if ( isAtStart && selectionStart ) {
-			onMultiSelect( { start: null, end: null } );
-		}
-
-		if ( ! isAtStart && selectionEnd !== uid ) {
-			onMultiSelect( { start: selectionAtStart, end: uid } );
-		}
-	}
-
-	onSelectionEnd() {
-		this.setState( { selectionAtStart: null } );
-	}
-
-	render() {
-		const { blocks, insertionPoint } = this.props;
-		const insertionPointIndex = blocks.indexOf( insertionPoint );
-		const blocksWithInsertionPoint = insertionPoint
-			? [
-				...blocks.slice( 0, insertionPointIndex + 1 ),
-				INSERTION_POINT_PLACEHOLDER,
-				...blocks.slice( insertionPointIndex + 1 ),
-			]
-			: blocks;
-		// Disable reason: Focus transfer between blocks and key events are handled
-		// by focused block element. Consider unhandled click bubbling as unselect.
-
-		/* eslint-disable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
-		return (
-			<div
-				role="region"
-				aria-label={ __( 'Visual Editor' ) }
-				className="editor-visual-editor"
-				onMouseDown={ this.onPress }
-				onTouchStart={ this.onPress }
-				onMouseUp={ this.onRelease }
-				onTouchEnd={ this.onRelease }
-			>
-				<PostTitle />
-				{ blocksWithInsertionPoint.map( ( uid ) => {
-					if ( uid === INSERTION_POINT_PLACEHOLDER ) {
-						return <div key={ INSERTION_POINT_PLACEHOLDER } className="editor-visual-editor__insertion-point" />;
-					}
-
-					return (
-						<VisualEditorBlock
-							key={ uid }
-							uid={ uid }
-							selectionStart={ this.state.selectionStart }
-							onSelectionStart={ () => this.onSelectionStart( uid ) }
-							onSelectionChange={ () => this.onSelectionChange( uid ) }
-							onSelectionEnd={ this.onSelectionEnd }
-						/>
-					);
-				} ) }
-				<Inserter position="top right" />
-			</div>
-		);
-		/* eslint-enable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
-	}
+export default function VisualEditor() {
+	return (
+		<div
+			role="region"
+			aria-label={ __( 'Visual Editor' ) }
+			className="editor-visual-editor"
+		>
+			<PostTitle />
+			<VisualEditorBlockList />
+			<Inserter position="top right" />
+		</div>
+	);
 }
-
-export default connect(
-	( state ) => ( {
-		blocks: getBlockUids( state ),
-		insertionPoint: getBlockInsertionPoint( state ),
-		selectionStart: getBlockSelectionStart( state ),
-		selectionEnd: getBlockSelectionEnd( state ),
-	} ),
-	( dispatch ) => ( {
-		clearSelectedBlock: () => dispatch( { type: 'CLEAR_SELECTED_BLOCK' } ),
-		onMultiSelect( { start, end } ) {
-			dispatch( { type: 'MULTI_SELECT', start, end } );
-		},
-	} ) )( VisualEditor );
